+++ tomcat/src/main/java/org/apache/coyote/http11/header/HttpHeaders.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.request;
+package org.apache.coyote.http11.header;
 
 import java.util.Map;
